requirements.txt
langgraph 
crewai 
openai 
streamlit 
faiss-cpu 
pypdf 
pymupdf 
tiktoken 
sqlalchemy
python-dotenv
PyPDF2